+++ packages/js-utils/src/Got/getKy.ts
@@ -1,68 +1,29 @@
-import { URL } from 'url'
-import { Time } from '@goatlab/js-utils'
-import { got } from 'got-cjs'
+import { Time } from '../Time'
+import Ky from './ky'
 import type {
-  Got,
-  AfterResponseHook,
+  KyInstance,
   BeforeErrorHook,
   BeforeRequestHook,
   BeforeRetryHook
-} from 'got-cjs'
-import { inspectAny } from './Got/inspectAny'
-import { GetGotOptions, GotRequestContext } from './Got/got.model'
+} from './ky'
+import { inspectAny } from './inspectAny'
+import { GetKyOptions, KyRequestContext } from './ky.model'
 
 /**
- * Returns instance of Got with "reasonable defaults":
+ * Returns instance of Ky with "reasonable defaults":
  *
  * 1. Error handler hook that prints helpful errors.
  * 2. Hooks that log start/end of request (optional, false by default).
  * 3. Reasonable defaults(tm), e.g non-infinite Timeout
- * 4. Preserves error stack traces (!) (experimental!)
  */
-export function getGot(opt: GetGotOptions = {}): Got {
+export function getKy(opt: GetKyOptions = {}): KyInstance {
   if (opt.debug) {
     opt.logStart = opt.logFinished = opt.logResponse = opt.logRequest = true
   }
 
-  return got.extend({
-    // Most-important is to set to anything non-empty (so, requests don't "hang" by default).
-    // Should be long enough to handle for slow responses from scaled cloud APIs in times of spikes
-    // Ideally should be LESS than default Request timeout in backend-lib (so, it has a chance to error
-    // before server times out with 503).
-    //
-    // UPD 2021-11-27
-    // There are 2 types/strategies for requests:
-    // 1. Optimized to get result no matter what. E.g in Cron jobs, where otherwise there'll be a job failure
-    // 2. Part of the Backend request, where we better retry quickly and fail on timeout before Backend aborts it with "503 Request timeout"
-    //
-    // Here it's hard to set the default timeout right for both use-cases.
-    // So, if it's important, you should override it according to your use-cases:
-    // - set it longer for Type 1 (e.g 120 seconds)
-    // - set it shorter for Type 2 (e.g 10/20 seconds)
-    // Please beware of default Retry strategy of Got:
-    // by default it will retry 2 times (after first try)
-    // First delay between tries will be ~1 second, then ~2 seconds
-    // Each retry it'll wait up to `timeout` (so, up to 60 seconds by default).
-    // So, for 3 tries it multiplies your timeout by 3 (+3 seconds between the tries).
-    // So, e.g 60 seconds timeout with 2 retries becomes up to 183 seconds.
-    // Which definitely doesn't fit into default "RequestTimeout"
-    timeout: {
-      response: 60_000
-    },
+  return Ky.extend({
+    timeout: 60_000,
     ...opt,
-    handlers: [
-      (options, next) => {
-        options.context = {
-          ...options.context,
-          started: Date.now(),
-          // This is to preserve original stack trace
-          // https://github.com/sindresorhus/got/blob/main/documentation/async-stack-traces.md
-          err: new Error('RequestError')
-        } as GotRequestContext
-
-        return next(options)
-      }
-    ],
     hooks: {
       ...opt.hooks,
       beforeError: [
@@ -81,15 +42,15 @@ export function getGot(opt: GetGotOptions = {}): Got {
         ...(opt.hooks?.beforeRetry || [])
       ],
       afterResponse: [
-        ...(opt.hooks?.afterResponse || []),
+        ...(opt.hooks?.afterResponse || [])
         // User hooks go BEFORE
-        gotAfterResponseHook(opt)
+        //keyAfterResponseHook(opt)
       ]
     }
   })
 }
 
-function gotErrorHook(opt: GetGotOptions = {}): BeforeErrorHook {
+function gotErrorHook(opt: GetKyOptions = {}): BeforeErrorHook {
   const { maxResponseLength = 10_000 } = opt
 
   return err => {
@@ -99,8 +60,9 @@ function gotErrorHook(opt: GetGotOptions = {}): BeforeErrorHook {
       return err
     }
 
-    const statusCode = err.response?.statusCode || 0
-    const { method, url, prefixUrl } = err.options
+    const statusCode = err.response?.status || 0
+    const { method, prefixUrl } = err.options
+    const url = err.request.url as any
 
     let shortUrl
 
@@ -109,19 +71,20 @@ function gotErrorHook(opt: GetGotOptions = {}): BeforeErrorHook {
       shortUrl = getShortUrl(
         opt,
         url instanceof URL ? url : new URL(url),
-        prefixUrl instanceof URL ? prefixUrl.toString() : prefixUrl
+        (prefixUrl as any) instanceof URL ? prefixUrl.toString() : prefixUrl
       )
     } catch (e) {
       console.error('Invalid URL:', url)
       shortUrl = url
     }
 
-    const context = err.request?.options?.context || {}
-    const { started, retryCount } = context as GotRequestContext
+    const context = err.options || {}
+    const { started, retryCount } = context as KyRequestContext
+
     if (!context) {
       console.error(
         'Context is not defined in the request options:',
-        err.request?.options
+        err.options
       )
       return err
     }
@@ -157,7 +120,8 @@ function gotErrorHook(opt: GetGotOptions = {}): BeforeErrorHook {
       .filter(Boolean)
       .join('\n')
 
-    const stack = (err.options.context as GotRequestContext)?.err?.stack
+    const stack = err.stack
+
     if (stack) {
       const originalStack = err.stack.split('\n')
       let originalStackIndex = originalStack.findIndex(line =>
@@ -187,16 +151,21 @@ function gotErrorHook(opt: GetGotOptions = {}): BeforeErrorHook {
   }
 }
 
-function gotBeforeRequestHook(opt: GetGotOptions): BeforeRequestHook {
-  return options => {
+function gotBeforeRequestHook(opt: GetKyOptions): BeforeRequestHook {
+  return (request, options) => {
+    if (!options['context']) {
+      options['context'] = {}
+      options['context']['started'] = new Date()
+    }
+
     // If options or options.context are not defined, log an error and exit early.
-    if (!options || !options.context) {
+    if (!options || !options['context']) {
       console.error('Unexpected options:', options)
       return
     }
 
     if (opt.logStart) {
-      const { retryCount } = options.context as GotRequestContext
+      const { retryCount } = options['context'] as KyRequestContext
 
       let shortUrl
 
@@ -204,14 +173,16 @@ function gotBeforeRequestHook(opt: GetGotOptions): BeforeRequestHook {
       try {
         shortUrl = getShortUrl(
           opt,
-          options.url instanceof URL ? options.url : new URL(options.url),
-          options.prefixUrl instanceof URL
+          (request.url as any) instanceof URL
+            ? (request.url as any)
+            : new URL(request.url),
+          (options.prefixUrl as any) instanceof URL
             ? options.prefixUrl.toString()
             : options.prefixUrl
         )
       } catch (e) {
-        console.error('Invalid URL:', options.url)
-        shortUrl = options.url
+        console.error('Invalid URL:', request.url)
+        shortUrl = request.url
       }
 
       console.log(
@@ -222,7 +193,7 @@ function gotBeforeRequestHook(opt: GetGotOptions): BeforeRequestHook {
     }
 
     if (opt.logRequest) {
-      const body = options.json || options.body
+      const body = request.body
 
       if (body) {
         console.log(body)
@@ -237,25 +208,26 @@ function gotBeforeRequestHook(opt: GetGotOptions): BeforeRequestHook {
  * @param opt - The options for the got HTTP client
  * @returns The before retry hook
  */
-function gotBeforeRetryHook(opt: GetGotOptions): BeforeRetryHook {
+function gotBeforeRetryHook(opt: GetKyOptions): BeforeRetryHook {
   const { maxResponseLength = 10_000 } = opt
 
-  return async (err, retryCount) => {
-    const statusCode = err?.response?.statusCode || 0
+  return async ({ error, request, retryCount, options }) => {
+    // const statusCode = request. || 0
 
-    // Skip hook for successful status codes
-    if (statusCode && statusCode < 300) {
-      return
-    }
+    // // Skip hook for successful status codes
+    // if (statusCode && statusCode < 300) {
+    //   return
+    // }
 
-    const { method, url, prefixUrl } = opt
+    const url = request.url
+    const { method, prefixUrl } = options
+
     let shortUrl
 
     // Safely construct shortUrl
     try {
-      const urlObject = url instanceof URL ? url : new URL(url)
-      const prefixUrlString =
-        prefixUrl instanceof URL ? prefixUrl.toString() : prefixUrl
+      const urlObject = new URL(url)
+      const prefixUrlString = prefixUrl
       shortUrl = getShortUrl(opt, urlObject, prefixUrlString)
     } catch (e) {
       console.error('Invalid URL:', url)
@@ -263,24 +235,26 @@ function gotBeforeRetryHook(opt: GetGotOptions): BeforeRetryHook {
     }
 
     // Ensure opt.context exists
-    if (!opt.context) {
-      opt.context = {}
+    if (!opt['context']) {
+      opt['context'] = {}
     }
-    const { started } = opt.context as GotRequestContext
-    opt.context = { ...opt.context, retryCount }
+
+    const { started } = opt['context'] as KyRequestContext
+
+    opt['context'] = { ...opt['context'], retryCount }
 
     // Construct body message
-    const body = err?.response?.body
-      ? inspectAny(err.response.body, {
+    const body = error.message
+      ? inspectAny(error, {
           maxLen: maxResponseLength,
           colors: false
         })
-      : err?.message
+      : error.message
 
     // Construct and log the warning message
     const messageParts = [
       ' <<',
-      statusCode,
+      //statusCode,
       method,
       shortUrl,
       retryCount && retryCount > 1
@@ -302,55 +276,54 @@ function gotBeforeRetryHook(opt: GetGotOptions): BeforeRetryHook {
  * @param opt - The options for the got HTTP client
  * @returns The after response hook
  */
-function gotAfterResponseHook(opt: GetGotOptions = {}): AfterResponseHook {
-  return resp => {
-    const success = resp.statusCode >= 200 && resp.statusCode < 400
+// function keyAfterResponseHook(opt: GetKyOptions = {}): AfterResponseHook {
+//   return (request, options, resp ) => {
+//     const success = resp.status >= 200 && resp.status < 400
 
-    // Errors are not logged here, as they're logged by gotErrorHook
-    if (opt.logFinished && success) {
-      const requestOptions = resp.request.options
+//     // Errors are not logged here, as they're logged by kyErrorHook
+//     if (opt.logFinished && success) {
+//       const requestOptions = options
+//       options.
+//       // Ensure context exists
+//       if (!requestOptions.context) {
+//         requestOptions.context = {}
+//       }
 
-      // Ensure context exists
-      if (!requestOptions.context) {
-        requestOptions.context = {}
-      }
+//       const { started, retryCount } = requestOptions.context as KyRequestContext
+//       const { url, prefixUrl, method } = requestOptions
+//       let shortUrl
 
-      const { started, retryCount } =
-        requestOptions.context as GotRequestContext
-      const { url, prefixUrl, method } = requestOptions
-      let shortUrl
+//       try {
+//         const urlObject = url instanceof URL ? url : new URL(url)
+//         const prefixUrlString =
+//           prefixUrl instanceof URL ? prefixUrl.toString() : prefixUrl
+//         shortUrl = getShortUrl(opt, urlObject, prefixUrlString)
+//       } catch (e) {
+//         console.error('Invalid URL:', url)
+//         shortUrl = url
+//       }
 
-      try {
-        const urlObject = url instanceof URL ? url : new URL(url)
-        const prefixUrlString =
-          prefixUrl instanceof URL ? prefixUrl.toString() : prefixUrl
-        shortUrl = getShortUrl(opt, urlObject, prefixUrlString)
-      } catch (e) {
-        console.error('Invalid URL:', url)
-        shortUrl = url
-      }
+//       const logParts = [
+//         ' <<',
+//         resp.status,
+//         method,
+//         shortUrl,
+//         retryCount && `(retry ${retryCount - 1})`,
+//         started && 'in ' + Time.since(started)
+//       ]
+//       const logMessage = logParts.filter(Boolean).join(' ')
 
-      const logParts = [
-        ' <<',
-        resp.statusCode,
-        method,
-        shortUrl,
-        retryCount && `(retry ${retryCount - 1})`,
-        started && 'in ' + Time.since(started)
-      ]
-      const logMessage = logParts.filter(Boolean).join(' ')
+//       console.log(logMessage)
+//     }
 
-      console.log(logMessage)
-    }
+//     // Error responses are not logged, cause they're included in Error message already
+//     if (opt.logResponse && success) {
+//       console.log(inspectAny(resp.body, { maxLen: opt.maxResponseLength }))
+//     }
 
-    // Error responses are not logged, cause they're included in Error message already
-    if (opt.logResponse && success) {
-      console.log(inspectAny(resp.body, { maxLen: opt.maxResponseLength }))
-    }
-
-    return resp
-  }
-}
+//     return resp
+//   }
+// }
 
 /**
  * Returns a shortened URL based on given options.
@@ -360,7 +333,11 @@ function gotAfterResponseHook(opt: GetGotOptions = {}): AfterResponseHook {
  * @param prefixUrl - The prefix URL to be removed if option is set
  * @returns The shortened URL as a string
  */
-function getShortUrl(opt: GetGotOptions, url: URL, prefixUrl?: string): string {
+export function getShortUrl(
+  opt: GetKyOptions,
+  url: URL,
+  prefixUrl?: string
+): string {
   // Copy the URL and redact password if it exists
   if (url.password) {
     const redactedUrl = new URL(url.toString())