+++ packages/js-utils/src/Got/inspectAny.ts
@@ -1,17 +1,14 @@
 import { inspect, InspectOptions } from 'util'
-import { Strings } from '@goatlab/js-utils'
-import type {
-  StringifyAnyOptions,
-  JsonStringifyFunction
-} from '@goatlab/js-utils'
+import { Strings } from '../Strings'
+import type { StringifyAnyOptions, JsonStringifyFunction } from '../index'
 
 export interface InspectAnyOptions
   extends StringifyAnyOptions,
     InspectOptions {}
 
 const INSPECT_OPT: InspectOptions = {
-  breakLength: 80, // default: ??
-  depth: 10 // default: 2
+  breakLength: 80,
+  depth: 10
 }
 
 /**
@@ -24,18 +21,6 @@ export const inspectAnyStringifyFn: JsonStringifyFunction = obj =>
  * Transforms ANY to human-readable string (via util.inspect mainly).
  * Safe (no error throwing).
  *
- * Correclty prints Errors, AppErrors, ErrorObjects: error.message + \n + inspect(error.data)
- *
- * Enforces max length (default to 10_000, pass 0 to skip it).
- *
- * Logs numbers as-is, e.g: `6`.
- * Logs strings as-is (without single quotes around, unlike default util.inspect behavior).
- * Otherwise - just uses util.inspect() with reasonable defaults.
- *
- * Returns 'empty_string' if empty string is passed.
- * Returns 'undefined' if undefined is passed (default util.inspect behavior).
- *
- * Based on `_stringifyAny` from `js-lib`, just replaced `JSON.stringify` with `util.inspect`.
  */
 export function inspectAny(obj: any, opt: InspectAnyOptions = {}): string {
   // Inspect handles functions better