+++ packages/node-backend/src/index.ts
@@ -1,0 +1,5 @@
+import { Cache } from './Cache'
+import type { Options } from 'keyv'
+
+export { Cache }
+export type { Options }