+++ packages/node-utils/src/Inspect.ts
@@ -1,9 +1,0 @@
-import { inspectAny, inspectAnyStringifyFn } from "./Got/inspectAny"
-
-class InspectClass {
-    anyStringifyFn = inspectAnyStringifyFn
-
-    any = inspectAny
-}
-
-export const Inspect = new InspectClass()