+++ packages/node-utils/src/streams/pipelineClose.ts
@@ -6,13 +6,13 @@ export function pipelineClose(
   readableDownstream: Readable,
   sourceReadable: Readable | undefined,
   streamDone: Promise<void> | undefined,
-  logger: CommonLogger,
+  logger: CommonLogger
 ): void {
   readableDownstream.push(null) // this closes the stream, so downstream Readable will receive `end` and won't write anything
 
   if (!sourceReadable) {
-    logger.warn(
-      `${name} sourceReadable is not provided, readable stream will not be stopped`,
+    logger.log(
+      `${name} sourceReadable is not provided, readable stream will not be stopped`
     )
   } else {
     logger.log(`${name} is calling readable.unpipe() to pause the stream`)
@@ -20,7 +20,7 @@ export function pipelineClose(
 
     if (!streamDone) {
       logger.log(
-        `${name} streamDone is not provided, will do readable.destroy right away`,
+        `${name} streamDone is not provided, will do readable.destroy right away`
       )
       sourceReadable.destroy()
     } else {