+++ packages/queue-core/src/test/const.ts
@@ -1,8 +1,9 @@
-import { readFileSync } from 'fs'
+import { existsSync, readFileSync, unlinkSync, writeFileSync } from 'fs'
 import { join, resolve } from 'path'
 
 export type GlobalTempData = {
   rabbitMQUrl?: string
+  kafkaUrl?: string
 }
 const tempDataFilePath = resolve(join(__dirname, '../../'), 'tempData.json')
 
@@ -10,4 +11,14 @@ export const getGlobalData = (): GlobalTempData => {
   const data = JSON.parse(readFileSync(tempDataFilePath, 'utf-8'))
 
   return data
+}
+
+export const writeGlobalData = (data: GlobalTempData) => {
+  writeFileSync(tempDataFilePath, JSON.stringify(data), 'utf-8')
+}
+
+export const cleanGlobalData = () => {
+  if (existsSync(tempDataFilePath)) {
+    unlinkSync(tempDataFilePath)
+  }
 }