+++ packages/fluent-firebase/CHANGELOG.md
@@ -1,5 +1,37 @@
 # 0.5.20
 
+## 0.7.24
+
+### Patch Changes
+
+- Updated dependencies
+  - @goatlab/js-utils@0.8.40
+  - @goatlab/fluent@0.7.39
+
+## 0.7.23
+
+### Patch Changes
+
+- Updated dependencies
+  - @goatlab/js-utils@0.8.39
+  - @goatlab/fluent@0.7.38
+
+## 0.7.22
+
+### Patch Changes
+
+- Updated dependencies
+  - @goatlab/js-utils@0.8.38
+  - @goatlab/fluent@0.7.37
+
+## 0.7.21
+
+### Patch Changes
+
+- Updated dependencies
+  - @goatlab/js-utils@0.8.37
+  - @goatlab/fluent@0.7.36
+
 ## 0.7.20
 
 ### Patch Changes