+++ packages/fluent-formio/CHANGELOG.md
@@ -1,5 +1,45 @@
 # 0.5.20
 
+## 0.6.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @goatlab/js-utils@0.8.40
+  - @goatlab/fluent@0.7.39
+
+## 0.6.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @goatlab/js-utils@0.8.39
+  - @goatlab/fluent@0.7.38
+
+## 0.6.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @goatlab/js-utils@0.8.38
+  - @goatlab/fluent@0.7.37
+
+## 0.6.11
+
+### Patch Changes
+
+- Remove getGot from node-utils
+- Updated dependencies
+  - @goatlab/js-utils@0.8.37
+  - @goatlab/fluent@0.7.36
+
+## 0.6.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @goatlab/fluent@0.7.35
+
 ## 0.6.9
 
 ### Patch Changes