+++ packages/fluent-loki/CHANGELOG.md
@@ -1,5 +1,54 @@
 # 0.5.20
 
+## 0.7.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @goatlab/js-utils@0.8.40
+  - @goatlab/fluent@0.7.39
+
+## 0.7.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @goatlab/js-utils@0.8.39
+  - @goatlab/fluent@0.7.38
+
+## 0.7.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @goatlab/js-utils@0.8.38
+  - @goatlab/fluent@0.7.37
+
+## 0.7.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @goatlab/tsconfig@0.0.12
+  - @goatlab/fluent@0.7.36
+  - @goatlab/js-utils@0.8.37
+
+## 0.7.6
+
+### Patch Changes
+
+- Remove getGot from node-utils
+- Updated dependencies
+  - @goatlab/js-utils@0.8.37
+  - @goatlab/fluent@0.7.36
+
+## 0.7.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @goatlab/fluent@0.7.35
+
 ## 0.7.4
 
 ### Patch Changes