+++ packages/node-utils/CHANGELOG.md
@@ -1,5 +1,61 @@
 # 0.5.20
 
+## 0.8.58
+
+### Patch Changes
+
+- Updated dependencies
+  - @goatlab/js-utils@0.8.40
+
+## 0.8.57
+
+### Patch Changes
+
+- Updated dependencies
+  - @goatlab/js-utils@0.8.39
+
+## 0.8.56
+
+### Patch Changes
+
+- Updated dependencies
+  - @goatlab/js-utils@0.8.38
+
+## 0.8.55
+
+### Patch Changes
+
+- Change ts config
+- Updated dependencies
+  - @goatlab/tsconfig@0.0.12
+  - @goatlab/js-utils@0.8.37
+
+## 0.8.54
+
+### Patch Changes
+
+- Remove getGot from node-utils
+- Updated dependencies
+  - @goatlab/js-utils@0.8.37
+
+## 0.8.53
+
+### Patch Changes
+
+- Add node-backend
+
+## 0.8.52
+
+### Patch Changes
+
+- Fix Ips
+
+## 0.8.51
+
+### Patch Changes
+
+- Create backend repo
+
 ## 0.8.50
 
 ### Patch Changes